backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
openai==1.30.1
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
httpx==0.27.0
fpdf2==2.7.9
python-multipart==0.0.9
